Core Functionalities Done, auth, menu, booking and ready to implement the front

+++ controllers/menu.controllers.ts
@@ -1,5 +1,6 @@
 import { type Request, type Response } from "express";
 import MenuItem from "../models/menu.model";
+import { ValidateToken } from "../utils/validateToken";
 
 export class MenuControllers {
   public constructor() {}
@@ -8,10 +9,9 @@ export class MenuControllers {
     try {
       const result = await MenuItem.find({});
 
-      res.status(200).json({ data: result });
+      res.status(200).json(result);
     } catch (err) {
-      console.error(err);
-      res.status(500).json({ message: "Internal Server Error" });
+      ValidateToken.catchTokenErrors(res, err as Error);
     }
   }
   
@@ -21,47 +21,56 @@ export class MenuControllers {
 
       const result = await MenuItem.findById(_id);
 
-      res.status(200).json({ data: result });
+      if(!result) {
+        res.status(404).json({message: "There is no item with this ID in the Menu"});
+      }
+
+      res.status(200).json(result);
     } catch (err) {
-      console.error(err);
-      res.status(500).json({ message: "Internal Server Error" });
+      ValidateToken.catchTokenErrors(res, err as Error);
     }
   }
 
   public async addMenuItem(req: Request, res: Response) {
     try {
+      const adminId = await ValidateToken.validateToken(req);
+      if(!adminId) return;
+
       const result = await MenuItem.create(req.body);
 
       res.status(201).json(result);
     } catch (err) {
-      console.error(err);
-      res.status(400).json({ message: "Bad Request" });
+      ValidateToken.catchTokenErrors(res, err as Error);
     }
   }
 
   public async deleteMenuItem(req: Request, res: Response) {
     try {
+      const adminId = await ValidateToken.validateToken(req);
+      if(!adminId) return;
+
       const { _id } = req.query;
       
       const result = await MenuItem.deleteOne({ _id });
 
-      res.status(200).json({data: result, message: "The Menu Item Has Been Deleted Successfully!"});
+      res.status(200).json(result);
     } catch (err) {
-      console.error(err);
-      res.status(400).json({ message: "Bad Request" });
+      ValidateToken.catchTokenErrors(res, err as Error);
     }
   }
   
   public async updateMenuItem(req: Request, res: Response) {
     try {
+      const adminId = await ValidateToken.validateToken(req);
+      if(!adminId) return;
+
       const { _id } = req.query;
       
       const result = await MenuItem.updateOne({ _id }, req.body);
 
-      res.status(200).json({data: result, message: "The Menu Item Has Been Updated Successfully!"});
+      res.status(200).json(result);
     } catch (err) {
-      console.error(err);
-      res.status(400).json({ message: "Bad Request" });
+      ValidateToken.catchTokenErrors(res, err as Error);
     }
   }
 }
